refactor: keep only usable constructs in the folder [sc-0]
