+++ packages/cli/src/constructs/check-group.ts
@@ -1,8 +1,6 @@
 import * as path from 'path'
 import * as glob from 'glob'
-import { Ref } from './ref'
-import { Session } from './project'
-import { Construct } from './construct'
+import { Construct, Session, Ref } from '../internals'
 import { BrowserCheck } from './browser-check'
 import { AlertChannel } from './alert-channel'
 import { EnvironmentVariable } from './environment-variable'
