+++ packages/cli/src/constructs/opsgenie-alert-channel.ts
@@ -1,5 +1,5 @@
 import { AlertChannel, AlertChannelProps } from './alert-channel'
-import { Session } from './project'
+import { Session } from '../internals'
 
 export enum OpsgeniePriority {
   P1 = 'P1',
