+++ packages/cli/src/internals/project.ts
