+++ packages/cli/src/services/checkly-config-loader.ts
@@ -2,8 +2,7 @@ import * as path from 'path'
 import { existsSync } from 'fs'
 import { loadJsFile, loadTsFile } from './util'
 import { CheckProps } from '../constructs/check'
-import { Session } from '../constructs'
-import { Construct } from '../constructs/construct'
+import { Session, Construct } from '../internals'
 
 export type CheckConfigDefaults = Pick<CheckProps, 'activated' | 'muted' | 'doubleCheck'
   | 'shouldFail' | 'runtimeId' | 'locations' | 'tags' | 'frequency' | 'environmentVariables'
